asdasd

+++ service/src/main/java/org/example/auction/AuctionServiceImpl.java
@@ -2,10 +2,12 @@ package org.example.auction;
 
 import jakarta.transaction.Transactional;
 import org.example.Auction;
+import org.example.exception.*;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Service
@@ -18,8 +20,14 @@ public class AuctionServiceImpl implements AuctionService {
     }
 
     @Override
-    public Auction createAuction(String name, LocalDateTime startsAt, String description, Double price) {
-        return auctionRepository.save(toAuction(name, startsAt, description, price));
+    public Auction createAuction(
+            String name,
+            LocalDateTime startsAt,
+            String description,
+            BigDecimal price,
+            int userId
+    ) {
+        return auctionRepository.save(toAuction(name, startsAt, description, price, userId));
     }
 
     @Override
@@ -28,29 +36,75 @@ public class AuctionServiceImpl implements AuctionService {
     }
 
     @Override
-    public Auction patchAuction(int id, String name, String description, Double price, LocalDateTime startsAt) {
+    public Auction patchAuction(int id, String name, String description, BigDecimal price, LocalDateTime startsAt) {
         var auction = auctionRepository.requireById(id);
+        var now = LocalDateTime.now();
+
+        if (auction.getStartsAt().isBefore(now)) {
+            throw new AuctionAlreadyStartedException(auction.getId());
+        }
+
         if (name != null) auction.setName(name);
         if (description != null) auction.setDescription(description);
-        if (price != null) auction.setPrice(price);
-        if (startsAt != null) auction.setStartsAtAndRenewEndsAt(startsAt);
+        if (price != null) auction.setStartPrice(price);
+        if (startsAt != null) {
+            if (startsAt.isBefore(now)) {
+                throw new StartsAtExpiredException(startsAt);
+            }
+            auction.setStartsAtAndRenewEndsAt(startsAt);
+        }
         return auction;
     }
 
     @Override
     public void deleteById(int id) {
         var auction = auctionRepository.requireById(id);
+
+        if (auction.getStartsAt().isBefore(LocalDateTime.now())) {
+            throw new AuctionAlreadyStartedException(auction.getId());
+        }
+
         auctionRepository.delete(auction);
     }
 
-    private Auction toAuction(String name, LocalDateTime startsAt, String description, Double price) {
+    @Override
+    public Auction bidOnAuction(int id, int winnerId, String winnerEmail, BigDecimal price) {
+        var auction = auctionRepository.requireById(id);
+        var now = LocalDateTime.now();
+
+        if (auction.getCurrentPrice().doubleValue() >= price.doubleValue()) {
+            throw new AuctionPriceException(auction.getId(), price.doubleValue());
+        }
+
+        if (now.isBefore(auction.getStartsAt())) {
+            throw new AuctionNotStartedException(auction.getId());
+        }
+
+        if (now.isAfter(auction.getEndsAt())) {
+            throw new AuctionAlreadyEndedException(auction.getId());
+        }
+
+        auction.setCurrentPrice(price);
+        auction.setWinnerEmail(winnerEmail);
+        auction.setWinnerId(winnerId);
+
+        return auctionRepository.save(auction);
+    }
+
+    @Override
+    public Auction requireById(int id) {
+        return auctionRepository.requireById(id);
+    }
+
+    private Auction toAuction(String name, LocalDateTime startsAt, String description, BigDecimal price, int userId) {
         var auction = new Auction();
-        auction.setActive(true);
         auction.setEndsAt(startsAt.plusHours(3));
         auction.setDescription(description);
-        auction.setPrice(price);
+        auction.setStartPrice(price);
         auction.setName(name);
         auction.setStartsAt(startsAt);
+        auction.setCurrentPrice(price);
+        auction.setOwnerId(userId);
         return auction;
     }
 }
